requirements.txt
mysql-connector-python==8.2.0
Pillow==10.1.0